0 errores

+++ Practica 2/plantillaSDL/HolaSDL/Mushroom.h
@@ -24,10 +24,10 @@ private:
 	Game* game = nullptr;		// puntero al juego
 
 	// MOVIMIENTO
-	Point2D<double> position;	// posicion actual en Point2D
+	Point2D<int> position;	// posicion actual en Point2D
 	Vector2D<int> direction;	// direccion de movimiento
 
-	Point2D<double> new_position;	// posicion que tendria si avanza en Point2D
+	Point2D<int> new_position;	// posicion que tendria si avanza en Point2D
 	SDL_Rect new_rect;				// rect que tendria si avanza en Point2D
 
 	SDL_Rect destRect;
@@ -35,7 +35,7 @@ private:
 	bool alive;
 
 public:
-	Mushroom(Game* g, Point2D<double> p);
+	Mushroom(Game* g, Point2D<int> p);
 
 	// -- update --
 	void update() override;
